move paper images download tests
