+++ tests/unit/gatys_ecker_bethge_2016/test_data.py
@@ -1,20 +1,6 @@
-import pytest
-
 import pystiche_papers.gatys_ecker_bethge_2016 as paper
-
-from tests.asserts import assert_image_downloads_correctly, assert_image_is_downloadable
-
-
-@pytest.mark.slow
-def test_images_smoke(subtests):
-    for name, image in paper.images():
-        with subtests.test(name=name):
-            assert_image_is_downloadable(image)
+from pystiche.data import DownloadableImageCollection
 
 
-@pytest.mark.large_download
-@pytest.mark.slow
-def test_images(subtests):
-    for name, image in paper.images():
-        with subtests.test(name=name):
-            assert_image_downloads_correctly(image)
+def test_images_smoke():
+    assert isinstance(paper.images(), DownloadableImageCollection)
